xmltvs

+++ src/main/java/com/connectto/guide/entity/IPTVChannelCountryTimezone.java
@@ -12,7 +12,7 @@ public class IPTVChannelCountryTimezone {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private int id;
+    private long id;
 
     @Column(name = "name")
     private String name;
@@ -55,11 +55,11 @@ public class IPTVChannelCountryTimezone {
      *##################################################################################################################
      */
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public IPTVChannelCountryTimezone setId(int id) {
+    public IPTVChannelCountryTimezone setId(long id) {
         this.id = id;
         return this;
     }
